fix typo
The current standard ebook text says "Nor forever". 

I looked this up because I was sure this was an OCR error and it should be "Not forever", but in both the 1928 edition that is the "official" source and in another 1930 release I checked, it is "Nor for ever" - adding a space. So adding a pull request.